# openapi.yaml
openapi: 3.0.3
info:
  title: Médiathèque API
  description: |
    Backend API for a Tunisian media platform with two main sections:
    - **Contenus à impact**: Videos, podcasts, reels, documentaries
    - **Tounesna**: Tourism photo gallery with watermarked previews
  version: 1.0.0
  contact:
    name: CnBees Team
    email: contact@mediatheque.tn
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.mediatheque.tn/api
    description: Production server
...
paths:
...
  /photos:
    get:
      tags: [Photos]
      summary: List photos with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: governorate
          in: query
          schema:
            type: string
        - name: landscapeType
          in: query
          schema:
            type: string
            enum: [sea, desert, mountain, village, oasis, forest, city, historical, other]
      responses:
        '200':
          description: List of photos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoList'
  /photos/{id}:
    get:
      tags: [Photos]
      summary: Get photo by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Photo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
  /photos/{id}/preview:
    get:
      tags: [Photos]
      summary: Get watermarked low-res preview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
  /admin/content/upload:
    post:
      tags: [Admin]
      summary: Upload new content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title, type]
              properties:
                file:
                  type: string
                  format: binary
                thumbnail:
                  type: string
                  format: binary
                title:
                  type: string
                type:
                  type: string
                  enum: [video, audio, reel, documentary]
                description:
                  type: string
      responses:
        '201':
          description: Content uploaded
  /admin/photos/upload:
    post:
      tags: [Admin]
      summary: Upload new photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [highRes, title, governorate, landscapeType, priceTND]
              properties:
                highRes:
                  type: string
                  format: binary
                lowRes:
                  type: string
                  format: binary
                title:
                  type: string
                governorate:
                  type: string
                landscapeType:
                  type: string
                priceTND:
                  type: number
      responses:
        '201':
          description: Photo uploaded
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Photo:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        governorate:
          type: string
        landscapeType:
          type: string
        priceTND:
          type: number
        previewUrl:
          type: string
    PhotoList:
      type: object
      properties:
        status:
          type: string
        results:
          type: integer
        total:
          type: integer
        data:
          type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/Photo'
...